backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
python-pptx>=1.0.0
github-copilot-sdk
